refactor: split MissionControl.tsx into sub-components (#129)
Extract 719-line MissionControl into:
- useMissionControl.ts (356 lines) — all state, refs, callbacks, effects
- MissionControlHub.tsx (63 lines) — center hub / ON STAGE agents
- MissionControlCluster.tsx (164 lines) — session cluster rendering
- MissionControl.tsx (240 lines) — thin wrapper with SVG + overlays
